back-end/api/healthy/views.py
from rest_framework import status
from rest_framework import viewsets
from rest_framework.response import Response
from rest_framework.permissions import AllowAny, IsAuthenticated


class ApiHealthyViewSet(viewsets.ViewSet):
    permission_classes = (IsAuthenticated,)

    def list(self, request):
        return Response(data={"message: healthy"}, status=status.HTTP_200_OK)
